fix_build

+++ dist-electron/ipc/license.ipc.js
@@ -1,5 +1,6 @@
 import { ipcMain } from 'electron';
 import { licenseService } from '../license/license.service.js';
+import { verifyPermission } from '../utils/permissions.js';
 export function registerLicenseHandlers() {
     // Initialize on startup
     licenseService.initialize();
@@ -7,9 +8,11 @@ export function registerLicenseHandlers() {
         return licenseService.getInternalState();
     });
     ipcMain.handle('license:activate', async (_, { key }) => {
+        verifyPermission('MANAGE_LICENSE');
         return await licenseService.activateLicense(key);
     });
     ipcMain.handle('license:delete', () => {
+        verifyPermission('MANAGE_LICENSE');
         return licenseService.deleteLicense();
     });
 }
