+++ electron/ipc/license.ipc.ts
@@ -1,5 +1,6 @@
 import { ipcMain } from 'electron';
 import { licenseService } from '../license/license.service.js';
+import { verifyPermission } from '../utils/permissions.js';
 
 export function registerLicenseHandlers() {
 
@@ -11,10 +12,12 @@ export function registerLicenseHandlers() {
     });
 
     ipcMain.handle('license:activate', async (_, { key }) => {
+        verifyPermission('MANAGE_LICENSE');
         return await licenseService.activateLicense(key);
     });
 
     ipcMain.handle('license:delete', () => {
+        verifyPermission('MANAGE_LICENSE');
         return licenseService.deleteLicense();
     });
 }
